requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
infoblox-netmri==3.8.0.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
